Fix legacy voice

+++ samples/batch-synthesis/csharp/BatchSynthesisSample/Program.cs
@@ -3,34 +3,17 @@
 // Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
 //
 
-
 // Your Speech resource key and region
 // This example requires environment variables named "SPEECH_KEY" and "SPEECH_REGION"
-string speechRegion;
-string speechKey;
+string speechRegion = Environment.GetEnvironmentVariable("SPEECH_REGION") 
+    ?? throw new ArgumentException("Please set SPEECH_REGION environment variable.");
+
+string speechKey = Environment.GetEnvironmentVariable("SPEECH_KEY") 
+    ?? throw new ArgumentException("Please set the SPEECH_KEY environment variable to set speech resource key.");
+
 string apiVersion = "2024-04-01";
-
-if (Environment.GetEnvironmentVariable("SPEECH_REGION") is string regionValue)
-{
-    speechRegion = regionValue;
-}
-else
-{
-    throw new ArgumentException($"Please set SPEECH_REGION environment variable.");
-}
-
-if (Environment.GetEnvironmentVariable("SPEECH_KEY") is string keyValue)
-{
-    speechKey = keyValue;
-}
-else
-{
-    throw new ArgumentException($"Please set the SPEECH_KEY environment variable to set speech resource key.");
-}
-
 var host = $"https://{speechRegion}.api.cognitive.microsoft.com";
-
-var sampleScript = await File.ReadAllTextAsync("Gatsby-chapter1.txt").ConfigureAwait(false);
+var sampleScript = await File.ReadAllTextAsync("../../Gatsby-chapter1.txt").ConfigureAwait(false);
 
 var synthesisClient = new BatchSynthesisClient(host, speechKey, apiVersion);
 
@@ -38,14 +21,10 @@ var synthesisClient = new BatchSynthesisClient(host, speechKey, apiVersion);
 var synthesisJobs = await synthesisClient.GetAllSynthesesAsync().ConfigureAwait(false);
 Console.WriteLine($"Found {synthesisJobs.Count()} jobs.");
 
-var newJobId = $"SimpleJob-{DateTime.Now.ToString("u").Replace(":", "-").Replace(" ", "-")}";
+var newJobId = $"SimpleJob-{DateTime.Now:u}".Replace(":", "-").Replace(" ", "-");
 
 // Create a new synthesis job with plain text
-await synthesisClient.CreateSynthesisAsync(
-   newJobId,
-   "AvaNeural",
-   sampleScript,
-   false).ConfigureAwait(false);
+await synthesisClient.CreateSynthesisAsync(newJobId, "en-US-AvaMultilingualNeural", sampleScript, false).ConfigureAwait(false);
 
 // Get a synthesis job.
 var synthesis = await synthesisClient.GetSynthesisAsync(newJobId).ConfigureAwait(false);
@@ -74,6 +53,15 @@ if (!string.IsNullOrEmpty(synthesis.Outputs?.Summary))
     Console.WriteLine(synthesis.Outputs.Summary);
 }
 
-// Delete a specific synthesis
-await synthesisClient.DeleteSynthesisAsync(newJobId);
-Console.WriteLine($"Deleted synthesis {newJobId}.");
+// Ask for confirmation before deleting the synthesis
+Console.Write("Do you want to delete the synthesis job? (y/n): ");
+var input = Console.ReadLine();
+if (input?.Trim().ToLower() == "y")
+{
+    await synthesisClient.DeleteSynthesisAsync(newJobId);
+    Console.WriteLine($"Deleted synthesis {newJobId}.");
+}
+else
+{
+    Console.WriteLine("Synthesis deletion cancelled.");
+}
